requirements.txt
streamlit>=1.32.0
langchain>=1.0.0
langchain-openai>=0.3.0
langchain-community>=0.4.0
langchain-core>=1.0.0
langchain-text-splitters>=0.3.0
PyPDF2>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
